// packages/shopify/generated/graphql.ts
export type QueryRoot = {
  __typename?: 'QueryRoot'
  /** List of the shop's articles. */
  articles: ArticleConnection
  /** Fetch a specific `Blog` by one of its unique attributes. */
  blog?: Maybe<Blog>
  /**
   * Find a blog by its handle.
   * @deprecated Use `blog` instead.
   */
  blogByHandle?: Maybe<Blog>
  /** List of the shop's blogs. */
  blogs: BlogConnection
  /**
   * Retrieve a cart by its ID. For more information, refer to
   * [Manage a cart with the Storefront API](https://shopify.dev/custom-storefronts/cart/manage).
   *
   */
  cart?: Maybe<Cart>
  /** Fetch a specific `Collection` by one of its unique attributes. */
  collection?: Maybe<Collection>
  /**
   * Find a collection by its handle.
   * @deprecated Use `collection` instead.
   */
  collectionByHandle?: Maybe<Collection>
  /** List of the shop’s collections. */
  collections: CollectionConnection
  /** Find a customer by its access token. */
  customer?: Maybe<Customer>
  /** Returns the localized experiences configured for the shop. */
  localization: Localization
  /**
   * List of the shop's locations that support in-store pickup.
   *
   * When sorting by distance, you must specify a location via the `near` argument.
   *
   */
  locations: LocationConnection
  /** A storefront menu. */
  menu?: Maybe<Menu>
  /** Fetch a specific Metaobject by one of its unique identifiers. */
  metaobject?: Maybe<Metaobject>
  /** All active metaobjects for the shop. */
  metaobjects: MetaobjectConnection
  /** Returns a specific node by ID. */
  node?: Maybe<Node>
  /** Returns the list of nodes with the given IDs. */
  nodes: Array<Maybe<Node>>
  /** Fetch a specific `Page` by one of its unique attributes. */
  page?: Maybe<Page>
  /**
   * Find a page by its handle.
   * @deprecated Use `page` instead.
   */
  pageByHandle?: Maybe<Page>
  /** List of the shop's pages. */
  pages: PageConnection
  /** Fetch a specific `Product` by one of its unique attributes. */
  product?: Maybe<Product>
  /**
   * Find a product by its handle.
   * @deprecated Use `product` instead.
   */
  productByHandle?: Maybe<Product>
  /**
   * Find recommended products related to a given `product_id`.
   * To learn more about how recommendations are generated, see
   * [*Showing product recommendations on product pages*](https://help.shopify.com/themes/development/recommended-products).
   *
   */
  productRecommendations?: Maybe<Array<Product>>
  /**
   * Tags added to products.
   * Additional access scope required: unauthenticated_read_product_tags.
   *
   */
  productTags: StringConnection
  /** List of product types for the shop's products that are published to your app. */
  productTypes: StringConnection
  /** List of the shop’s products. */
  products: ProductConnection
  /** The list of public Storefront API versions, including supported, release candidate and unstable versions. */
  publicApiVersions: Array<ApiVersion>
  /** The shop associated with the storefront access token. */
  shop: Shop
  /** A list of redirects for a shop. */
  urlRedirects: UrlRedirectConnection
}
// ...
export type ShopPolicy = Node & {
  __typename?: 'ShopPolicy'
  /** Policy text, maximum size of 64kb. */
  body: Scalars['String']
  /** Policy’s handle. */
  handle: Scalars['String']
  /** A globally-unique identifier. */
  id: Scalars['ID']
  /** Policy’s title. */
  title: Scalars['String']
  /** Public URL to the policy. */
  url: Scalars['URL']
}
// ...
export type QueryRootInformationFragment = ({
  __typename?: 'QueryRoot'
  collections: { __typename?: 'CollectionConnection' } & {
    ' $fragmentRefs'?: {
      CollectionsBasicInformationFragment: CollectionsBasicInformationFragment
    }
  }
  shop: { __typename?: 'Shop' } & {
    ' $fragmentRefs'?: {
      ShopBasicInformationFragment: ShopBasicInformationFragment
    }
  }
} & {
  ' $fragmentRefs'?: { MenusInformationFragment: MenusInformationFragment }
}) & { ' $fragmentName'?: 'QueryRootInformationFragment' }

export type ShopBasicInformationFragment = {
  __typename?: 'Shop'
  name: string
  description?: string | null
} & { ' $fragmentName'?: 'ShopBasicInformationFragment' }

export type ShopPolicyInformationFragment = {
  __typename?: 'ShopPolicy'
  body: string
  handle: string
  id: string
  title: string
  url: string
} & { ' $fragmentName'?: 'ShopPolicyInformationFragment' }
// ...
export type PoliciesQueryVariables = Exact<{
  country: CountryCode
}>

export type PoliciesQuery = {
  __typename?: 'QueryRoot'
  shop: {
    __typename?: 'Shop'
    privacyPolicy?:
      | ({ __typename?: 'ShopPolicy' } & {
          ' $fragmentRefs'?: {
            ShopPolicyInformationFragment: ShopPolicyInformationFragment
          }
        })
      | null
    refundPolicy?:
      | ({ __typename?: 'ShopPolicy' } & {
          ' $fragmentRefs'?: {
            ShopPolicyInformationFragment: ShopPolicyInformationFragment
          }
        })
      | null
    shippingPolicy?:
      | ({ __typename?: 'ShopPolicy' } & {
          ' $fragmentRefs'?: {
            ShopPolicyInformationFragment: ShopPolicyInformationFragment
          }
        })
      | null
    termsOfService?:
      | ({ __typename?: 'ShopPolicy' } & {
          ' $fragmentRefs'?: {
            ShopPolicyInformationFragment: ShopPolicyInformationFragment
          }
        })
      | null
  }
} & {
  ' $fragmentRefs'?: {
    QueryRootInformationFragment: QueryRootInformationFragment
  }
}
// ...
export const ShopBasicInformationFragmentDoc = {
  kind: 'Document',
  definitions: [
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'ShopBasicInformation' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'Shop' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          { kind: 'Field', name: { kind: 'Name', value: 'name' } },
          { kind: 'Field', name: { kind: 'Name', value: 'description' } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<ShopBasicInformationFragment, unknown>
export const QueryRootInformationFragmentDoc = {
  kind: 'Document',
  definitions: [
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'QueryRootInformation' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'QueryRoot' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'FragmentSpread',
            name: { kind: 'Name', value: 'MenusInformation' },
          },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'collections' },
            arguments: [
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'first' },
                value: { kind: 'IntValue', value: '192' },
              },
            ],
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'FragmentSpread',
                  name: { kind: 'Name', value: 'CollectionsBasicInformation' },
                },
              ],
            },
          },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'shop' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'FragmentSpread',
                  name: { kind: 'Name', value: 'ShopBasicInformation' },
                },
              ],
            },
          },
        ],
      },
    },
    ...MenusInformationFragmentDoc.definitions,
    ...CollectionsBasicInformationFragmentDoc.definitions,
    ...ShopBasicInformationFragmentDoc.definitions,
  ],
} as unknown as DocumentNode<QueryRootInformationFragment, unknown>
export const ShopPolicyInformationFragmentDoc = {
  kind: 'Document',
  definitions: [
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'ShopPolicyInformation' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'ShopPolicy' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          { kind: 'Field', name: { kind: 'Name', value: 'body' } },
          { kind: 'Field', name: { kind: 'Name', value: 'handle' } },
          { kind: 'Field', name: { kind: 'Name', value: 'id' } },
          { kind: 'Field', name: { kind: 'Name', value: 'title' } },
          { kind: 'Field', name: { kind: 'Name', value: 'url' } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<ShopPolicyInformationFragment, unknown>
// ...
export const PoliciesDocument = {
  kind: 'Document',
  definitions: [
    {
      kind: 'OperationDefinition',
      operation: 'query',
      name: { kind: 'Name', value: 'Policies' },
      variableDefinitions: [
        {
          kind: 'VariableDefinition',
          variable: {
            kind: 'Variable',
            name: { kind: 'Name', value: 'country' },
          },
          type: {
            kind: 'NonNullType',
            type: {
              kind: 'NamedType',
              name: { kind: 'Name', value: 'CountryCode' },
            },
          },
        },
      ],
      directives: [
        {
          kind: 'Directive',
          name: { kind: 'Name', value: 'inContext' },
          arguments: [
            {
              kind: 'Argument',
              name: { kind: 'Name', value: 'country' },
              value: {
                kind: 'Variable',
                name: { kind: 'Name', value: 'country' },
              },
            },
          ],
        },
      ],
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'FragmentSpread',
            name: { kind: 'Name', value: 'QueryRootInformation' },
          },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'shop' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'privacyPolicy' },
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'FragmentSpread',
                        name: { kind: 'Name', value: 'ShopPolicyInformation' },
                      },
                    ],
                  },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'refundPolicy' },
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'FragmentSpread',
                        name: { kind: 'Name', value: 'ShopPolicyInformation' },
                      },
                    ],
                  },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'shippingPolicy' },
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'FragmentSpread',
                        name: { kind: 'Name', value: 'ShopPolicyInformation' },
                      },
                    ],
                  },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'termsOfService' },
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'FragmentSpread',
                        name: { kind: 'Name', value: 'ShopPolicyInformation' },
                      },
                    ],
                  },
                },
              ],
            },
          },
        ],
      },
    },
    ...QueryRootInformationFragmentDoc.definitions,
    ...ShopPolicyInformationFragmentDoc.definitions,
  ],
} as unknown as DocumentNode<PoliciesQuery, PoliciesQueryVariables>